remove exception package and move ConfigurationException to cfg package
