+++ src/main/java/com/mongodb/hibernate/NotYetImplementedException.java
@@ -14,24 +14,13 @@
  * limitations under the License.
  */
 
-package com.mongodb.hibernate.exception;
+package com.mongodb.hibernate;
 
-public class ConfigurationException extends RuntimeException {
+public class NotYetImplementedException extends RuntimeException {
 
-    private final String property;
+    public NotYetImplementedException() {}
 
-    public ConfigurationException(String property, String message) {
+    public NotYetImplementedException(String message) {
         super(message);
-        this.property = property;
-    }
-
-    public ConfigurationException(String property, String message, Throwable cause) {
-        super(message, cause);
-        this.property = property;
-    }
-
-    @Override
-    public String getMessage() {
-        return String.format("Invalid '%s' configuration: %s", property, super.getMessage());
     }
 }
