+++ src/main/java/com/mongodb/hibernate/jdbc/MongoConnectionProvider.java
@@ -23,9 +23,9 @@ import com.mongodb.MongoClientSettings;
 import com.mongodb.MongoCredential;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
+import com.mongodb.hibernate.NotYetImplementedException;
+import com.mongodb.hibernate.cfg.ConfigurationException;
 import com.mongodb.hibernate.cfg.ConfigurationHelper;
-import com.mongodb.hibernate.exception.ConfigurationException;
-import com.mongodb.hibernate.exception.NotYetImplementedException;
 import java.sql.Connection;
 import java.util.Map;
 import org.hibernate.engine.jdbc.connections.spi.ConnectionProvider;
